+++ Source Code/backend/competitions/utils.py
@@ -10,6 +10,11 @@ import itertools
 from collections import defaultdict
 from .models import Appearance, Grade, MediaFile, Result
 from datetime import datetime, date, time, timedelta
+import uuid
+from django.conf import settings
+from django.http import JsonResponse
+from django.core.mail import send_mail
+from rest_framework import status
 
 
 def calculate_start_time(base_time: time, length: timedelta) -> time:
@@ -191,3 +196,36 @@ def generate_grades(appearance):
     }
     
     return grades_json
+
+
+def send_judge_invite(request):
+    email = request.data.get('email')
+
+    if not email:
+        return JsonResponse(
+            {"detail": "Email is required"},
+            status=status.HTTP_400_BAD_REQUEST
+        )
+
+    token = uuid.uuid4()
+    base_url = settings.FRONTEND_URL.rstrip('/')
+    invite_link = f"{base_url}/?invitedJudge=True"
+
+    send_mail(
+        subject="Judge Registration Invitation",
+        message=(
+            "Hello,\n\n"
+            "You have been invited to register as a judge on Dance Arena.\n"
+            "Please use the link below to register:\n\n"
+            f"{invite_link}\n\n"
+            "Best regards,\n"
+            "Dance Arena Team"
+        ),
+        from_email=settings.DEFAULT_FROM_EMAIL,
+        recipient_list=[email],
+    )
+
+    return JsonResponse(
+        {"detail": "Invitation email sent"},
+        status=status.HTTP_200_OK
+    )
